move some of the logic out of views.py -- it is getting busy
